Add files via upload
The algorithm I created  opens the file that contains the "allow_list.txt"  I assigned as a string to the variable  import_file. It opens the file that contains the "allow_list.txt", to read the file I used .read( ) method that converts it back to strings so that I can read them. I used the .split() method to convert the ip_addresses string into a list.I setup a for loop header to iterate through the remove_list that contains all the ip_addresses that should be removed.I applied the .remove() method to the ip_addresses list and remove the IP addresses identified in the loop variable element.i updated the file with the revised list of ip_addresses by  converting the ip_addresses list back into a string using the .join() method. I applied the  .join() method to combine  all items in an iterable into a string. the string "\n" separate the elements in the file by placing them on a new line